Add "aria-current" attr to menu and breadcrumb
Add `[aria-current="page"]` to the menu item in "menu.html" that has
isActive=true.

Also add `[aria-current="location"]` to the current page in the bread
crumbs menu found in "breadcrumb.html". If `linkActiveItem` is set to
`false` the `<span>` wrapping the active item is not removed, but kept
to hold the "aria-current" attribute.

Screen readers use the "aria-current" attribute to indicate that a page
in a navigation list is the current page the user is visiting.

In "breadcrumb.html" an "aria-label" attribute is set on the "<nav>"
element so that a screen reader can communicate a menu name as a screen
reader user is navigating. The text used in the "aria-label" is
configured with a new param in `getBreadcrumbMenu()"` named
`navigationAriaLabel`.